Include default variant that doesn't use ag
